feat: 두 종류의 가짜 스케줄 바 컴포넌트를 하나의 컴포넌트로 통합
