--- frontend/src/components/team_calendar/ScheduleBarsIndicator/ScheduleBarsIndicator.styled.ts
@@ -1,15 +1,0 @@
-import { styled } from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  position: absolute;
-
-  width: 100%;
-  height: 100%;
-`;
-
-export const CalendarRow = styled.div`
-  position: relative;
-  flex-grow: 1;
-`;
